pyproject.toml

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "appleease"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.0",
    "cryptography>=46.0.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

# expose your Click group as a console script
[project.scripts]
appleease = "appleease.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
